Moved hpc_apps under src/testsuite/applications
