+++ src/testsuite/applications/hpc_apps/kokkos/kokkos_parser_common.py
