+++ src/testsuite/applications/hpc_apps/quicksilver/quicksilver_diff_patch
